Fixes for typing and add loading circle for network scans

+++ backend_py/src/services/lights/light.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 @dataclass
 class Light:
-    intensity: int
+    intensity: float
     pin: int
     controller_index: int
     controller_name: str
